Revert "ooojj"
This reverts commit 9be52d19a6e6a35968f2657035afafe74aa60d77.

+++ Proyecto_1_Grupo/Program.cs
@@ -12,8 +12,8 @@ namespace Proyecto_1_Grupo
         {
             public string modelo;
             public int anio;
-            public int kmActual;
-            public int kmService;
+            public long kmActual;
+            public long kmService;
             public string color;
             public string duenio;
         }
